vexpress: move some code to core/arch/arm/kernel
Some code that was previously in plat-vexpress is moved to the kernel
directory and enabled by CFG_GENERIC_BOOT=y so that it may be used by
other platforms.
Tested on QEMU and FVP, with and without pager.

Tested-by: Jerome Forissier <jerome.forissier@linaro.org> (QEMU, FVP)
