+++ core/arch/arm/plat-vexpress/conf.mk
@@ -41,6 +41,7 @@ CFG_GIC := y
 CFG_HWSUPP_MEM_PERM_PXN := y
 CFG_WITH_STACK_CANARIES := y
 CFG_PM_STUBS := y
+CFG_GENERIC_BOOT := y
 
 ifeq ($(PLATFORM_FLAVOR),juno)
 CFG_CRYPTO_SHA256_ARM32_CE ?= $(CFG_ARM32_core)
